chore: fixed report table
